dynamic resource loading

+++ source/stringutils.cpp
@@ -12,7 +12,7 @@ std::vector<std::string> StringUtils::split(std::string& str, char split) {
 
 std::string StringUtils::join(std::vector<std::string>& strings, std::string join) {
     std::string joined;
-    for (int i = 0; i < strings.size(); i++) {
+    for (size_t i = 0; i < strings.size(); i++) {
         joined += strings[i];
         if (i != strings.size() - 1) {
             joined += join;
